move formatDaterange to utils
